+++ cmd/norobod/norobo.go
@@ -1,16 +1,16 @@
 package main
 
 import (
-	"encoding/csv"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"net/http"
 	"os"
-	"strings"
 	"sync"
 	"time"
 
+	"github.com/dgnorton/norobo"
+	"github.com/dgnorton/norobo/filter"
 	"github.com/dgnorton/norobo/hayes"
 )
 
@@ -121,90 +121,20 @@ func (h *webHandler) serveCalls(w http.ResponseWriter, r *http.Request) {
 	w.Write(b)
 }
 
-// Call represents the current in-progress call.
-type Call struct {
-	*hayes.Call
-	FilterResult *FilterResult
-}
-
-// CallEntry represents a completed (ended) call.
-type CallEntry struct {
-	Time   time.Time `json:"time"`
-	Name   string    `json:"name"`
-	Number string    `json:"number"`
-	Action string    `json:"action"`
-	Filter string    `json:"filter"`
-	Rule   string    `json:"rule"`
-}
-
-// CallLog represents a list of completed (ended) calls.
-type CallLog struct {
-	Calls []*CallEntry `json:"calls"`
-}
-
-func (l *CallLog) LastTime() time.Time {
-	return l.Calls[len(l.Calls)-1].Time
-}
-
-func LoadCallLog(filename string) (*CallLog, error) {
-	f, err := os.Open(filename)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			return nil, err
-		}
-		return &CallLog{
-			Calls: make([]*CallEntry, 0),
-		}, nil
-	}
-
-	records, err := csv.NewReader(f).ReadAll()
-	if err != nil {
-		return nil, err
-	}
-
-	calls := &CallLog{
-		Calls: make([]*CallEntry, 0, len(records)),
-	}
-
-	for _, r := range records {
-		if len(r) != 6 {
-			return nil, fmt.Errorf("expected 6 fields but got %d: %s", len(r), strings.Join(r, ","))
-		}
-
-		t, err := time.Parse(time.RFC3339Nano, r[0])
-		if err != nil {
-			return nil, err
-		}
-
-		call := &CallEntry{
-			Time:   t,
-			Name:   r[1],
-			Number: r[2],
-			Action: r[3],
-			Filter: r[4],
-			Rule:   r[5],
-		}
-
-		calls.Calls = append(calls.Calls, call)
-	}
-
-	return calls, nil
-}
-
 type callHandler struct {
 	modem          *hayes.Modem
-	filters        Filters
+	filters        norobo.Filters
 	callLogFile    string
 	mu             sync.RWMutex
-	callLog        *CallLog
+	callLog        *norobo.CallLog
 	callLogChanged chan struct{}
 }
 
 func newCallHandler(m *hayes.Modem, blockFile, allowFile, callLogFile string) *callHandler {
-	filters := Filters{}
+	filters := norobo.Filters{}
 
 	if blockFile != "" {
-		block, err := LoadFilterFile(blockFile, Block)
+		block, err := filter.LoadFilterFile(blockFile, norobo.Block)
 		if err != nil {
 			panic(err)
 		}
@@ -212,14 +142,14 @@ func newCallHandler(m *hayes.Modem, blockFile, allowFile, callLogFile string) *c
 	}
 
 	if allowFile != "" {
-		allow, err := LoadFilterFile(allowFile, Allow)
+		allow, err := filter.LoadFilterFile(allowFile, norobo.Allow)
 		if err != nil {
 			panic(err)
 		}
 		filters = append(filters, allow)
 	}
 
-	callLog, err := LoadCallLog(callLogFile)
+	callLog, err := norobo.LoadCallLog(callLogFile)
 	if err != nil {
 		panic(err)
 	}
@@ -236,17 +166,17 @@ func newCallHandler(m *hayes.Modem, blockFile, allowFile, callLogFile string) *c
 }
 
 func (h *callHandler) Handle(c *hayes.Call) {
-	call := &Call{Call: c}
+	call := &norobo.Call{Call: c}
 
 	call.FilterResult = h.filters.Run(call)
-	if call.FilterResult.Action == Block {
+	if call.FilterResult.Action == norobo.Block {
 		call.Block()
 	}
 
 	h.log(call)
 }
 
-func (h *callHandler) CallLog() *CallLog {
+func (h *callHandler) CallLog() *norobo.CallLog {
 	h.mu.RLock()
 	defer h.mu.RUnlock()
 	return h.callLog
@@ -278,7 +208,7 @@ func (h *callHandler) CallLogChanged(after time.Time) chan struct{} {
 	return changedCh
 }
 
-func (h *callHandler) log(c *Call) {
+func (h *callHandler) log(c *norobo.Call) {
 	f, err := os.OpenFile(h.callLogFile, os.O_CREATE|os.O_APPEND|os.O_WRONLY, 0770)
 	if err != nil {
 		println(err)
@@ -289,7 +219,7 @@ func (h *callHandler) log(c *Call) {
 	msg := fmt.Sprintf(`"%s","%s","%s","%s","%s","%s"%s`, c.Time.Format(time.RFC3339Nano), c.Name, c.Number, r.Action, r.FilterDescription(), r.Description, "\n")
 
 	h.mu.Lock()
-	call := &CallEntry{
+	call := &norobo.CallEntry{
 		Time:   c.Time,
 		Name:   c.Name,
 		Number: c.Number,
