+++ filter/filter_test.go
@@ -1,4 +1,4 @@
-package main
+package filter
 
 import (
 	"testing"
